♻️ Refactor: checkLogin hook 구현
